moves layout under common
